ScotlandJS Demos
The code I used for the demonstrations in my 2015 ScotlandJS talk "A
brief history of synthesis".

This code is extracted from a more complicated build process which
if included, only distracts from the aim of the code - which is to teach
these principles of synthesis.

There's also some hacky UI code to switch synth presets...